Updated PaneHeader example to render a more complex action menu example
