+++ lib/Pane/stories/Pane.stories.js
@@ -6,7 +6,7 @@ import readme from '../readme.md';
 import BasicUsage from './BasicUsage';
 
 import PaneHeaderReadme from '../../PaneHeader/readme.md';
-import PaneHeader from './PaneHeader';
+import PaneHeaderBasicUsage from '../../PaneHeader/stories/BasicUsage';
 
 import PaneBackLinkReadme from '../../PaneBackLink/readme.md';
 import PaneBackLinkBasicUsage from '../../PaneBackLink/stories/BasicUsage';
@@ -18,7 +18,7 @@ import PaneMenuReadme from '../../PaneMenu/readme.md';
 
 storiesOf('Pane', module)
   .add('Basic Usage', withReadme(readme, () => <BasicUsage />))
-  .add('PaneHeader', withReadme(PaneHeaderReadme, () => <PaneHeader />))
+  .add('PaneHeader', withReadme(PaneHeaderReadme, () => <PaneHeaderBasicUsage />))
   .add('PaneMenu', withReadme(PaneMenuReadme, () => <PaneHeader />))
   .add('PaneBackLink', withReadme(PaneBackLinkReadme, () => <PaneBackLinkBasicUsage />))
   .add('PaneCloseLink', withReadme(PaneCloseLinkReadme, () => <PaneCloseLinkBasicUsage />));
